requirements.txt
fastapi
uvicorn
openai
feedparser
python-dotenv
azure-functions
